move config files to config folder; move tests to test folders
